+++ ovp_search/tests/test_views.py
@@ -405,6 +405,15 @@ class UserSearchTestCase(TestCase):
     self.assertEqual(str(response.data["results"][0]["profile"]["full_name"]), "user one")
     self.assertEqual(str(response.data["results"][1]["profile"]["full_name"]), "user two")
 
+    response = self.client.get(reverse("search-users-list") + "?cause=OR,1,2", format="json")
+    self.assertEqual(len(response.data["results"]), 2)
+    self.assertEqual(str(response.data["results"][0]["profile"]["full_name"]), "user one")
+    self.assertEqual(str(response.data["results"][1]["profile"]["full_name"]), "user two")
+
+    response = self.client.get(reverse("search-users-list") + "?cause=AND,1,2", format="json")
+    self.assertEqual(len(response.data["results"]), 1)
+    self.assertEqual(str(response.data["results"][0]["profile"]["full_name"]), "user one")
+
 
   def test_skills_filter(self):
     """